+++ src/lib/route/data/index.ts
@@ -2,4 +2,4 @@ export { LabelFunction, RouteData, ReadonlyRouteData } from './route-data-types'
 export { RouteDataInterface } from './route-data.interface';
 export { RouteDataRegistry } from './route-data-registry';
 export { RouteDataRegistryInitializer } from './route-data-registry-initializer';
-export { extractRouteDataFromRouterConfig, findMatchingRoutePath, getChildren } from './utils';
+export { extractRouteDataFromRouterConfig, findMatchingRoutePath, getChildren } from '../utils';
