+++ src/lib/route/data/route-data-registry-initializer.spec.ts
